refact: separate parse_rediction*() functions
